feat(relatorio-chamado-repository): adiciona metodo para buscar chamados por tecnico

+++ src/main/java/dev/joaorooliveira/help_desk/projection/ChamadoPorCategoriaProjection.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public interface ChamadoPorCategoriaProjection {
 
     String getCategoria();
-    BigDecimal total();
+    Long total();
 
 }
