+++ src/main/java/dev/joaorooliveira/help_desk/projection/ChamadoPorStatusProjection.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 public interface ChamadoPorStatusProjection {
 
     StatusTipo status();
-    BigDecimal total();
+    Long total();
 
 }
